Move DisclosureControllable out from the material page scope
Because we use it in the search header as well. Now it's in its own
folder on the components level.
